tests-requirements.txt
black
isort
pylint
pytest
umap-learn
anthropic
cohere
openai
azure-ai-inference
mistral-ai
numpy>=1.26,<2.0
transformers
bm25s
